normalize url

+++ ir_query_engine/retrieve_match_models/tf_idf_feature/unittests/test_tfidf_model.py
@@ -45,10 +45,11 @@ class TfIdfModelTestCase(unittest.TestCase):
 
     def tearDown(self):
         # cleanup
-        os.remove(self.test_dict_file_path)
-        os.remove(self.test_md_file_path)
-        os.remove(self.test_q_simmx_file_path)
-        os.remove(self.test_a_simmx_file_path)
+        if os.path.isfile(self.test_dict_file_path):
+            os.remove(self.test_dict_file_path)
+            os.remove(self.test_md_file_path)
+            os.remove(self.test_q_simmx_file_path)
+            os.remove(self.test_a_simmx_file_path)
 
     @patch('ir_query_engine.retrieve_match_models.tf_idf_feature.tfidf_model.get_md_path')
     @patch('ir_query_engine.retrieve_match_models.tf_idf_feature.tfidf_model.get_dict_path')
@@ -190,6 +191,14 @@ class TfIdfModelTestCase(unittest.TestCase):
                 model_struct.get_tfidf_vec(compare_docs[idx])
             )
             self.assertAlmostEqual(sim[1], expected_sim)
+
+    def test_process(self):
+        raw_doc = 'You can find the website of Matternet at http'
+        self.assertEqual(TfIdfModelStruct.pre_process_doc_tf_idf(raw_doc),
+                         [u'you', u'can', u'find', u'the', u'websit', u'of', u'matternet', u'at', u'http'])
+        raw_doc = 'What do you like to eat?'
+        self.assertEqual(TfIdfModelStruct.pre_process_doc_tf_idf(raw_doc),
+                         [u'what', u'do', u'you', u'like', u'to', u'eat'])
 
 
 if __name__ == '__main__':
